random improvements

+++ overlay/src/pages/overtime_and_sudden_death.rs
@@ -30,12 +30,12 @@ impl PageRenderer {
                 if time < 1f32 {
                     (
                         (0f32, -200f32).interpolate_linear(1f32 - time),
-                        (ALPHA_MIN, ALPHA_MAX).interpolate_exponential_end(time),
+                        (ALPHA_MIN, ALPHA_MAX).interpolate_exponential_end(time) as u8,
                     )
                 } else {
                     (
                         (0f32, -200f32).interpolate_linear(0f32),
-                        (ALPHA_MIN, ALPHA_MAX).interpolate_exponential_end(1f32),
+                        (ALPHA_MIN, ALPHA_MAX).interpolate_exponential_end(1f32) as u8,
                     )
                 }
             } else if self.last_snapshot_timeout != TimeoutSnapshot::None {
@@ -52,19 +52,19 @@ impl PageRenderer {
                     self.last_snapshot_timeout = TimeoutSnapshot::None;
                     (
                         (0f32, -200f32).interpolate_linear(1f32),
-                        (ALPHA_MIN, ALPHA_MAX).interpolate_exponential_end(0f32),
+                        (ALPHA_MIN, ALPHA_MAX).interpolate_exponential_end(0f32) as u8,
                     )
                 } else {
                     (
                         (0f32, -200f32).interpolate_linear(time),
-                        (ALPHA_MIN, ALPHA_MAX).interpolate_exponential_end(1f32 - time),
+                        (ALPHA_MIN, ALPHA_MAX).interpolate_exponential_end(1f32 - time) as u8,
                     )
                 }
             } else {
                 // return any values when both are None, cause we won't be redering anyways
                 (
                     (0f32, -200f32).interpolate_linear(0f32),
-                    (ALPHA_MIN, ALPHA_MAX).interpolate_exponential_end(1f32),
+                    (ALPHA_MIN, ALPHA_MAX).interpolate_exponential_end(1f32) as u8,
                 )
             };
 
@@ -77,7 +77,7 @@ impl PageRenderer {
                     self.textures.referee_timout_graphic,
                     timeout_offset + 380f32,
                     35f32,
-                    Color::from_rgba(255, 255, 255, timeout_alpha_offset as u8)
+                    Color::from_rgba(255, 255, 255, timeout_alpha_offset)
                 );
                 draw_text_both_ex!(
                     "REFEREE",
@@ -86,13 +86,13 @@ impl PageRenderer {
                     TextParams {
                         font: self.textures.font,
                         font_size: 20,
-                        color: Color::from_rgba(0, 0, 0, timeout_alpha_offset as u8),
+                        color: Color::from_rgba(0, 0, 0, timeout_alpha_offset),
                         ..Default::default()
                     },
                     TextParams {
                         font: self.textures.font,
                         font_size: 20,
-                        color: Color::from_rgba(255, 255, 255, timeout_alpha_offset as u8),
+                        color: Color::from_rgba(255, 255, 255, timeout_alpha_offset),
                         ..Default::default()
                     }
                 );
@@ -103,13 +103,13 @@ impl PageRenderer {
                     TextParams {
                         font: self.textures.font,
                         font_size: 20,
-                        color: Color::from_rgba(0, 0, 0, timeout_alpha_offset as u8),
+                        color: Color::from_rgba(0, 0, 0, timeout_alpha_offset),
                         ..Default::default()
                     },
                     TextParams {
                         font: self.textures.font,
                         font_size: 20,
-                        color: Color::from_rgba(255, 255, 255, timeout_alpha_offset as u8),
+                        color: Color::from_rgba(255, 255, 255, timeout_alpha_offset),
                         ..Default::default()
                     }
                 );
@@ -119,7 +119,7 @@ impl PageRenderer {
                     self.textures.penalty_graphic,
                     timeout_offset + 380f32,
                     35f32,
-                    Color::from_rgba(255, 255, 255, timeout_alpha_offset as u8)
+                    Color::from_rgba(255, 255, 255, timeout_alpha_offset)
                 );
                 draw_text_both_ex!(
                     "PENALTY",
@@ -128,13 +128,13 @@ impl PageRenderer {
                     TextParams {
                         font: self.textures.font,
                         font_size: 20,
-                        color: Color::from_rgba(0, 0, 0, timeout_alpha_offset as u8),
+                        color: Color::from_rgba(0, 0, 0, timeout_alpha_offset),
                         ..Default::default()
                     },
                     TextParams {
                         font: self.textures.font,
                         font_size: 20,
-                        color: Color::from_rgba(255, 255, 255, timeout_alpha_offset as u8),
+                        color: Color::from_rgba(255, 255, 255, timeout_alpha_offset),
                         ..Default::default()
                     }
                 );
@@ -145,13 +145,13 @@ impl PageRenderer {
                     TextParams {
                         font: self.textures.font,
                         font_size: 20,
-                        color: Color::from_rgba(0, 0, 0, timeout_alpha_offset as u8),
+                        color: Color::from_rgba(0, 0, 0, timeout_alpha_offset),
                         ..Default::default()
                     },
                     TextParams {
                         font: self.textures.font,
                         font_size: 20,
-                        color: Color::from_rgba(255, 255, 255, timeout_alpha_offset as u8),
+                        color: Color::from_rgba(255, 255, 255, timeout_alpha_offset),
                         ..Default::default()
                     }
                 );
